Update Navbar to handle responsive menu
Addition of a responsive menu feature to the Navbar component which improves the mobile user experience. This feature includes a 'handleClick' function to close the menu after a selection is made, and a 'showMenu' state for opening and closing the side menu. Social links with clickable icons were also added to enhance the menu.